Bring the full RPCS3 runtime to Wasm and WebGPU
Compile the existing RPCS3 emulator sources for Wasm32/pthreads, replace fixed host aliases with sparse guest mappings, add interpreter fallbacks, and provide Web-safe host shims for unavailable native facilities.\n\nAdd a WebGPU RSX backend and browser packet translator that preserve authentic shader, vertex, texture, depth, blend, clear, and flip state. Add complete-runtime and original-source GoogleTest harnesses, differential frame tooling, and USB WebKit device runners.\n\nSerialize hardware GPU acceptance tests so packet-loss assertions measure a single runtime rather than contention among eight independent GPU processes. Resolve the Emscripten SDK relative to the user cache rather than a machine-specific absolute path.\n\nValidation: full runtime and Wasm unit targets rebuild with Emscripten 6.0.8; npm check passes; default Playwright lane passes 3/3; hardware Chrome lane passes 6 enabled tests with 2 opt-in captures skipped; attached iPadOS 26.5.2 passes 114/114 compiled RPCS3 tests and renders three advancing cube frames with zero dropped packets.

+++ rpcs3/Emu/Cell/Modules/cellDmuxPamf.cpp
@@ -126,7 +126,7 @@ bool dmux_pamf_base::output_queue::push(const access_unit_chunk& au_chunk, const
 bool dmux_pamf_base::output_queue::prepare_next_au(u32 au_max_size)
 {
 	// LLE always checks the distance between the end of the buffer and the back pointer, even if the back pointer is wrapped around and there are unconsumed access units behind it
-	if (std::to_address(buffer.end()) - back < au_max_size)
+	if (static_cast<u32>(std::to_address(buffer.end()) - back) < au_max_size)
 	{
 		// Can't wrap the back pointer around again as long as there are unconsumed access units behind it
 		if (wrap_pos != buffer.data())
